make the `Continuation` trait more general; remove the old `Depend`
