+++ vest/src/regular/mod.rs
@@ -23,9 +23,6 @@ pub mod and_then;
 pub mod clone;
 /// Depend combinator
 pub mod depend;
-// pub mod dependent;
-/// Same as Depend but uses a Continuation trait instead of Fn
-pub mod depend_cont;
 /// Disjointness for OrdChoice combinator
 pub mod disjoint;
 /// Error combinator
